Changing the way we get EMA
I changed the way we are getting EMA
Now, we know if we are in an uptrend or in a downtrend looking at EMA sign

NN is not working properly (predicting NaN)
I still don't really get how it works

I changed the way we are calculating 'percentage'. If we use it that way, we know when we are in a short uptrend or downtrend. This might not be a good idea. Do we wan't to know the short trend, or the overall trend ?
For overall trend, we should not use a percentage, but just check lastEMA value. lastEMA>0 = uptrend      lastEMA<0 = downtrend